Create README.md
Set a word count target and a timer. Built for writers and people with deadlines.

When you run the program it will tell you the targeted WPM you will need based on current word count and how much time is left.

This uses spacebar presses to count words. This will get you 99% of the way there. Adjust the word count based on how often you find you delete spacebar presses.

For me, this is about 50 for every 500 words typed. YMMV.